requirements.txt
certifi==2024.8.30
charset-normalizer==2.1.1
idna==3.9
requests==2.28.1
urllib3==1.26.20
python-dotenv==1.0.1